requirements.txt
PyQt5
yt-dlp
openai-whisper
pytest
python-docx
moviepy
